refactor: rename create dir to create-post dir
