+++ src/app/create-post/CreateClientPage.tsx
@@ -1,6 +1,6 @@
 'use client'
 import React from 'react'
-import styles from '@/app/create/page.module.scss'
+import styles from '@/app/create-post/page.module.scss'
 import classNames from 'classnames/bind'
 
 const cn = classNames.bind(styles)
